// hp_app_angular/src/app/model/character-dto.ts
import { Wand } from './wand';





export interface CharacterDto{
    name: string;
    image: string;
    house: string;
    hogwartsStaff: boolean;
    dateOfBirth: string;
    actor: string;
    alive: boolean;
    wand: Wand;
    ancestry: string;
    eyeColour: string;
    hairColor: string;
    hogwartsStudent: string;
    patronus: string;
    species: string;
    yearOfBirth: string;

}